+++ src/plugins/score-plugin-scenario/Scenario/Settings/ScenarioSettingsView.hpp
@@ -36,7 +36,7 @@ public:
       W_SIGNAL(DefaultEditorChanged, arg_1);
 
   void zoomChanged(int arg_1) W_SIGNAL(zoomChanged, arg_1);
-  void SlotHeightChanged(qreal arg_1) W_SIGNAL(SlotHeightChanged, arg_1);
+  void SlotHeightChanged(int arg_1) W_SIGNAL(SlotHeightChanged, arg_1);
   void DefaultDurationChanged(const TimeVal& t)
       W_SIGNAL(DefaultDurationChanged, t);
